+++ towerhttp/compress.go
@@ -4,7 +4,7 @@ import (
 	"io"
 )
 
-type Compressor interface {
+type Compression interface {
 	// ContentEncoding returns the value of the Content-Encoding header. If empty, the content-encoding header will be
 	// set by the http framework.
 	ContentEncoding() string
@@ -14,11 +14,11 @@ type Compressor interface {
 	StreamCompress(origin io.Reader) (io.Reader, error)
 }
 
-var _ Compressor = (*NoopCompressor)(nil)
+var _ Compression = (*NoCompression)(nil)
 
-// NoopCompressor is a compressor that does nothing. Basically it's an Uncompressed operation.
-type NoopCompressor struct{}
+// NoCompression is a compressor that does nothing. Basically it's an Uncompressed operation.
+type NoCompression struct{}
 
-func (n NoopCompressor) StreamCompress(origin io.Reader) (io.Reader, error) { return origin, nil }
-func (n NoopCompressor) ContentEncoding() string                            { return "" }
-func (n NoopCompressor) Compress(bytes []byte) ([]byte, error)              { return bytes, nil }
+func (n NoCompression) StreamCompress(origin io.Reader) (io.Reader, error) { return origin, nil }
+func (n NoCompression) ContentEncoding() string                            { return "" }
+func (n NoCompression) Compress(bytes []byte) ([]byte, error)              { return bytes, nil }